PublicDashboards: refactor service
