+++ pkg/services/publicdashboards/publicdashboard.go
@@ -14,8 +14,7 @@ import (
 
 //go:generate mockery --name Service --structname FakePublicDashboardService --inpackage --filename public_dashboard_service_mock.go
 type Service interface {
-	AccessTokenExists(ctx context.Context, accessToken string) (bool, error)
-	BuildAnonymousUser(ctx context.Context, dashboard *models.Dashboard) *user.SignedInUser
+	EnabledPublicDashboardExistsByAccessToken(ctx context.Context, accessToken string) (bool, error)
 	GetAnnotations(ctx context.Context, reqDTO AnnotationsQueryDTO, accessToken string) ([]AnnotationEvent, error)
 	GetDashboard(ctx context.Context, dashboardUid string) (*models.Dashboard, error)
 	GetMetricRequest(ctx context.Context, dashboard *models.Dashboard, publicDashboard *PublicDashboard, panelId int64, reqDTO PublicDashboardQueryDTO) (dtos.MetricRequest, error)
@@ -30,7 +29,8 @@ type Service interface {
 
 //go:generate mockery --name Store --structname FakePublicDashboardStore --inpackage --filename public_dashboard_store_mock.go
 type Store interface {
-	AccessTokenExists(ctx context.Context, accessToken string) (bool, error)
+	EnabledPublicDashboardExistsByAccessToken(ctx context.Context, accessToken string) (bool, error)
+	EnabledPublicDashboardExistsByDashboardUid(ctx context.Context, dashboardUid string) (bool, error)
 	GenerateNewPublicDashboardUid(ctx context.Context) (string, error)
 	GetDashboard(ctx context.Context, dashboardUid string) (*models.Dashboard, error)
 	GenerateNewPublicDashboardAccessToken(ctx context.Context) (string, error)
@@ -39,7 +39,6 @@ type Store interface {
 	GetPublicDashboardConfig(ctx context.Context, orgId int64, dashboardUid string) (*PublicDashboard, error)
 	GetPublicDashboardOrgId(ctx context.Context, accessToken string) (int64, error)
 	ListPublicDashboards(ctx context.Context, orgId int64) ([]PublicDashboardListResponse, error)
-	PublicDashboardEnabled(ctx context.Context, dashboardUid string) (bool, error)
 	SavePublicDashboardConfig(ctx context.Context, cmd SavePublicDashboardConfigCommand) error
 	UpdatePublicDashboardConfig(ctx context.Context, cmd SavePublicDashboardConfigCommand) error
 }
