+++ pkg/services/publicdashboards/service/service_test.go
@@ -3,7 +3,6 @@ package service
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"testing"
 	"time"
@@ -14,25 +13,19 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/grafana/pkg/components/simplejson"
-	dashboard2 "github.com/grafana/grafana/pkg/coremodel/dashboard"
 	"github.com/grafana/grafana/pkg/infra/db"
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/services/accesscontrol"
-	"github.com/grafana/grafana/pkg/services/annotations"
-	"github.com/grafana/grafana/pkg/services/annotations/annotationsimpl"
 	"github.com/grafana/grafana/pkg/services/dashboards"
 	dashboardsDB "github.com/grafana/grafana/pkg/services/dashboards/database"
 	"github.com/grafana/grafana/pkg/services/featuremgmt"
 	. "github.com/grafana/grafana/pkg/services/publicdashboards"
 	"github.com/grafana/grafana/pkg/services/publicdashboards/database"
-	"github.com/grafana/grafana/pkg/services/publicdashboards/internal"
 	. "github.com/grafana/grafana/pkg/services/publicdashboards/models"
 	"github.com/grafana/grafana/pkg/services/serviceaccounts/tests"
-	"github.com/grafana/grafana/pkg/services/sqlstore"
 	"github.com/grafana/grafana/pkg/services/tag/tagimpl"
 	"github.com/grafana/grafana/pkg/services/user"
-	"github.com/grafana/grafana/pkg/setting"
 	"github.com/grafana/grafana/pkg/tsdb/intervalv2"
 )
 
@@ -43,289 +36,6 @@ var SignedInUser = &user.SignedInUser{UserID: 1234, Login: "user@login.com"}
 
 func TestLogPrefix(t *testing.T) {
 	assert.Equal(t, LogPrefix, "publicdashboards.service")
-}
-
-func TestGetAnnotations(t *testing.T) {
-	t.Run("will build anonymous user with correct permissions to get annotations", func(t *testing.T) {
-		sqlStore := sqlstore.InitTestDB(t)
-		config := setting.NewCfg()
-		tagService := tagimpl.ProvideService(sqlStore, sqlStore.Cfg)
-		annotationsRepo := annotationsimpl.ProvideService(sqlStore, config, tagService)
-		fakeStore := FakePublicDashboardStore{}
-		service := &PublicDashboardServiceImpl{
-			log:             log.New("test.logger"),
-			store:           &fakeStore,
-			AnnotationsRepo: annotationsRepo,
-		}
-		fakeStore.On("GetPublicDashboard", mock.Anything, mock.AnythingOfType("string")).
-			Return(&PublicDashboard{Uid: "uid1", IsEnabled: true}, models.NewDashboard("dash1"), nil)
-		reqDTO := AnnotationsQueryDTO{
-			From: 1,
-			To:   2,
-		}
-		dash := models.NewDashboard("testDashboard")
-
-		items, _ := service.GetAnnotations(context.Background(), reqDTO, "abc123")
-		anonUser := service.BuildAnonymousUser(context.Background(), dash)
-
-		assert.Equal(t, "dashboards:*", anonUser.Permissions[0]["dashboards:read"][0])
-		assert.Len(t, items, 0)
-	})
-
-	t.Run("Test events from tag queries overwrite built-in annotation queries and duplicate events are not returned", func(t *testing.T) {
-		dash := models.NewDashboard("test")
-		color := "red"
-		name := "annoName"
-		grafanaAnnotation := DashAnnotation{
-			Datasource: CreateDatasource("grafana", "grafana"),
-			Enable:     true,
-			Name:       &name,
-			IconColor:  &color,
-			Target: &dashboard2.AnnotationTarget{
-				Limit:    100,
-				MatchAny: false,
-				Tags:     nil,
-				Type:     "dashboard",
-			},
-			Type: "dashboard",
-		}
-		grafanaTagAnnotation := DashAnnotation{
-			Datasource: CreateDatasource("grafana", "grafana"),
-			Enable:     true,
-			Name:       &name,
-			IconColor:  &color,
-			Target: &dashboard2.AnnotationTarget{
-				Limit:    100,
-				MatchAny: false,
-				Tags:     []string{"tag1"},
-				Type:     "tags",
-			},
-		}
-		annos := []DashAnnotation{grafanaAnnotation, grafanaTagAnnotation}
-		dashboard := AddAnnotationsToDashboard(t, dash, annos)
-
-		annotationsRepo := annotations.FakeAnnotationsRepo{}
-		fakeStore := FakePublicDashboardStore{}
-		service := &PublicDashboardServiceImpl{
-			log:             log.New("test.logger"),
-			store:           &fakeStore,
-			AnnotationsRepo: &annotationsRepo,
-		}
-		pubdash := &PublicDashboard{Uid: "uid1", IsEnabled: true, OrgId: 1, DashboardUid: dashboard.Uid}
-		fakeStore.On("GetPublicDashboard", mock.Anything, mock.AnythingOfType("string")).Return(pubdash, dashboard, nil)
-		annotationsRepo.On("Find", mock.Anything, mock.Anything).Return([]*annotations.ItemDTO{
-			{
-				Id:          1,
-				DashboardId: 1,
-				PanelId:     1,
-				Tags:        []string{"tag1"},
-				TimeEnd:     2,
-				Time:        2,
-				Text:        "text",
-			},
-		}, nil).Maybe()
-
-		items, err := service.GetAnnotations(context.Background(), AnnotationsQueryDTO{}, "abc123")
-
-		expected := AnnotationEvent{
-			Id:          1,
-			DashboardId: 1,
-			PanelId:     0,
-			Tags:        []string{"tag1"},
-			IsRegion:    false,
-			Text:        "text",
-			Color:       color,
-			Time:        2,
-			TimeEnd:     2,
-			Source:      grafanaTagAnnotation,
-		}
-		require.NoError(t, err)
-		assert.Len(t, items, 1)
-		assert.Equal(t, expected, items[0])
-	})
-
-	t.Run("Test panelId set to zero when annotation event is for a tags query", func(t *testing.T) {
-		dash := models.NewDashboard("test")
-		color := "red"
-		name := "annoName"
-		grafanaAnnotation := DashAnnotation{
-			Datasource: CreateDatasource("grafana", "grafana"),
-			Enable:     true,
-			Name:       &name,
-			IconColor:  &color,
-			Target: &dashboard2.AnnotationTarget{
-				Limit:    100,
-				MatchAny: false,
-				Tags:     []string{"tag1"},
-				Type:     "tags",
-			},
-		}
-		annos := []DashAnnotation{grafanaAnnotation}
-		dashboard := AddAnnotationsToDashboard(t, dash, annos)
-
-		annotationsRepo := annotations.FakeAnnotationsRepo{}
-		fakeStore := FakePublicDashboardStore{}
-		service := &PublicDashboardServiceImpl{
-			log:             log.New("test.logger"),
-			store:           &fakeStore,
-			AnnotationsRepo: &annotationsRepo,
-		}
-		pubdash := &PublicDashboard{Uid: "uid1", IsEnabled: true, OrgId: 1, DashboardUid: dashboard.Uid}
-		fakeStore.On("GetPublicDashboard", mock.Anything, mock.AnythingOfType("string")).Return(pubdash, dashboard, nil)
-		annotationsRepo.On("Find", mock.Anything, mock.Anything).Return([]*annotations.ItemDTO{
-			{
-				Id:          1,
-				DashboardId: 1,
-				PanelId:     1,
-				Tags:        []string{},
-				TimeEnd:     1,
-				Time:        2,
-				Text:        "text",
-			},
-		}, nil).Maybe()
-
-		items, err := service.GetAnnotations(context.Background(), AnnotationsQueryDTO{}, "abc123")
-
-		expected := AnnotationEvent{
-			Id:          1,
-			DashboardId: 1,
-			PanelId:     0,
-			Tags:        []string{},
-			IsRegion:    true,
-			Text:        "text",
-			Color:       color,
-			Time:        2,
-			TimeEnd:     1,
-			Source:      grafanaAnnotation,
-		}
-		require.NoError(t, err)
-		assert.Len(t, items, 1)
-		assert.Equal(t, expected, items[0])
-	})
-
-	t.Run("Test can get grafana annotations and will skip annotation queries and disabled annotations", func(t *testing.T) {
-		dash := models.NewDashboard("test")
-		color := "red"
-		name := "annoName"
-		disabledGrafanaAnnotation := DashAnnotation{
-			Datasource: CreateDatasource("grafana", "grafana"),
-			Enable:     false,
-			Name:       &name,
-			IconColor:  &color,
-		}
-		grafanaAnnotation := DashAnnotation{
-			Datasource: CreateDatasource("grafana", "grafana"),
-			Enable:     true,
-			Name:       &name,
-			IconColor:  &color,
-			Target: &dashboard2.AnnotationTarget{
-				Limit:    100,
-				MatchAny: true,
-				Tags:     nil,
-				Type:     "dashboard",
-			},
-			Type: "dashboard",
-		}
-		queryAnnotation := DashAnnotation{
-			Datasource: CreateDatasource("prometheus", "abc123"),
-			Enable:     true,
-			Name:       &name,
-		}
-		annos := []DashAnnotation{grafanaAnnotation, queryAnnotation, disabledGrafanaAnnotation}
-		dashboard := AddAnnotationsToDashboard(t, dash, annos)
-
-		annotationsRepo := annotations.FakeAnnotationsRepo{}
-		fakeStore := FakePublicDashboardStore{}
-		service := &PublicDashboardServiceImpl{
-			log:             log.New("test.logger"),
-			store:           &fakeStore,
-			AnnotationsRepo: &annotationsRepo,
-		}
-		pubdash := &PublicDashboard{Uid: "uid1", IsEnabled: true, OrgId: 1, DashboardUid: dashboard.Uid}
-		fakeStore.On("GetPublicDashboard", mock.Anything, mock.AnythingOfType("string")).Return(pubdash, dashboard, nil)
-		annotationsRepo.On("Find", mock.Anything, mock.Anything).Return([]*annotations.ItemDTO{
-			{
-				Id:          1,
-				DashboardId: 1,
-				PanelId:     1,
-				Tags:        []string{},
-				TimeEnd:     1,
-				Time:        2,
-				Text:        "text",
-			},
-		}, nil).Maybe()
-
-		items, err := service.GetAnnotations(context.Background(), AnnotationsQueryDTO{}, "abc123")
-
-		expected := AnnotationEvent{
-			Id:          1,
-			DashboardId: 1,
-			PanelId:     1,
-			Tags:        []string{},
-			IsRegion:    true,
-			Text:        "text",
-			Color:       color,
-			Time:        2,
-			TimeEnd:     1,
-			Source:      grafanaAnnotation,
-		}
-		require.NoError(t, err)
-		assert.Len(t, items, 1)
-		assert.Equal(t, expected, items[0])
-	})
-
-	t.Run("test will return nothing when dashboard has no annotations", func(t *testing.T) {
-		annotationsRepo := annotations.FakeAnnotationsRepo{}
-		fakeStore := FakePublicDashboardStore{}
-		service := &PublicDashboardServiceImpl{
-			log:             log.New("test.logger"),
-			store:           &fakeStore,
-			AnnotationsRepo: &annotationsRepo,
-		}
-		dashboard := models.NewDashboard("dashWithNoAnnotations")
-		pubdash := &PublicDashboard{Uid: "uid1", IsEnabled: true, OrgId: 1, DashboardUid: dashboard.Uid}
-		fakeStore.On("GetPublicDashboard", mock.Anything, mock.AnythingOfType("string")).Return(pubdash, dashboard, nil)
-
-		items, err := service.GetAnnotations(context.Background(), AnnotationsQueryDTO{}, "abc123")
-
-		require.NoError(t, err)
-		assert.Empty(t, items)
-	})
-
-	t.Run("test will error when annotations repo returns an error", func(t *testing.T) {
-		annotationsRepo := annotations.FakeAnnotationsRepo{}
-		fakeStore := FakePublicDashboardStore{}
-		service := &PublicDashboardServiceImpl{
-			log:             log.New("test.logger"),
-			store:           &fakeStore,
-			AnnotationsRepo: &annotationsRepo,
-		}
-		dash := models.NewDashboard("test")
-		color := "red"
-		name := "annoName"
-		grafanaAnnotation := DashAnnotation{
-			Datasource: CreateDatasource("grafana", "grafana"),
-			Enable:     true,
-			Name:       &name,
-			IconColor:  &color,
-			Target: &dashboard2.AnnotationTarget{
-				Limit:    100,
-				MatchAny: false,
-				Tags:     []string{"tag1"},
-				Type:     "tags",
-			},
-		}
-		annos := []DashAnnotation{grafanaAnnotation}
-		dash = AddAnnotationsToDashboard(t, dash, annos)
-		pubdash := &PublicDashboard{Uid: "uid1", IsEnabled: true, OrgId: 1, DashboardUid: dash.Uid}
-		fakeStore.On("GetPublicDashboard", mock.Anything, mock.AnythingOfType("string")).Return(pubdash, dash, nil)
-		annotationsRepo.On("Find", mock.Anything, mock.Anything).Return(nil, errors.New("failed")).Maybe()
-
-		items, err := service.GetAnnotations(context.Background(), AnnotationsQueryDTO{}, "abc123")
-
-		require.Error(t, err)
-		require.Nil(t, items)
-	})
 }
 
 func TestGetPublicDashboard(t *testing.T) {
@@ -652,342 +362,6 @@ func TestUpdatePublicDashboard(t *testing.T) {
 		require.NoError(t, err)
 
 		assert.Equal(t, &TimeSettings{}, updatedPubdash.TimeSettings)
-	})
-}
-
-func TestBuildAnonymousUser(t *testing.T) {
-	sqlStore := db.InitTestDB(t)
-	dashboardStore := dashboardsDB.ProvideDashboardStore(sqlStore, sqlStore.Cfg, featuremgmt.WithFeatures(), tagimpl.ProvideService(sqlStore, sqlStore.Cfg))
-	dashboard := insertTestDashboard(t, dashboardStore, "testDashie", 1, 0, true, []map[string]interface{}{}, nil)
-	publicdashboardStore := database.ProvideStore(sqlStore)
-	service := &PublicDashboardServiceImpl{
-		log:   log.New("test.logger"),
-		store: publicdashboardStore,
-	}
-
-	t.Run("will add datasource read and query permissions to user for each datasource in dashboard", func(t *testing.T) {
-		user := service.BuildAnonymousUser(context.Background(), dashboard)
-
-		require.Equal(t, dashboard.OrgId, user.OrgID)
-		require.Equal(t, "datasources:uid:ds1", user.Permissions[user.OrgID]["datasources:query"][0])
-		require.Equal(t, "datasources:uid:ds3", user.Permissions[user.OrgID]["datasources:query"][1])
-		require.Equal(t, "datasources:uid:ds1", user.Permissions[user.OrgID]["datasources:read"][0])
-		require.Equal(t, "datasources:uid:ds3", user.Permissions[user.OrgID]["datasources:read"][1])
-	})
-	t.Run("will add dashboard and annotation permissions needed for getting annotations", func(t *testing.T) {
-		user := service.BuildAnonymousUser(context.Background(), dashboard)
-
-		require.Equal(t, dashboard.OrgId, user.OrgID)
-		require.Equal(t, "annotations:type:dashboard", user.Permissions[user.OrgID]["annotations:read"][0])
-		require.Equal(t, "dashboards:*", user.Permissions[user.OrgID]["dashboards:read"][0])
-	})
-}
-
-func TestGetMetricRequest(t *testing.T) {
-	sqlStore := db.InitTestDB(t)
-	dashboardStore := dashboardsDB.ProvideDashboardStore(sqlStore, sqlStore.Cfg, featuremgmt.WithFeatures(), tagimpl.ProvideService(sqlStore, sqlStore.Cfg))
-	publicdashboardStore := database.ProvideStore(sqlStore)
-	dashboard := insertTestDashboard(t, dashboardStore, "testDashie", 1, 0, true, []map[string]interface{}{}, nil)
-	publicDashboard := &PublicDashboard{
-		Uid:          "1",
-		DashboardUid: dashboard.Uid,
-		IsEnabled:    true,
-		AccessToken:  "abc123",
-	}
-	service := &PublicDashboardServiceImpl{
-		log:                log.New("test.logger"),
-		store:              publicdashboardStore,
-		intervalCalculator: intervalv2.NewCalculator(),
-	}
-
-	t.Run("will return an error when validation fails", func(t *testing.T) {
-		publicDashboardQueryDTO := PublicDashboardQueryDTO{
-			IntervalMs:    int64(-1),
-			MaxDataPoints: int64(-1),
-		}
-
-		_, err := service.GetMetricRequest(context.Background(), dashboard, publicDashboard, 1, publicDashboardQueryDTO)
-
-		require.Error(t, err)
-	})
-
-	t.Run("will not return an error when validation succeeds", func(t *testing.T) {
-		publicDashboardQueryDTO := PublicDashboardQueryDTO{
-			IntervalMs:    int64(1),
-			MaxDataPoints: int64(1),
-		}
-		from, to := internal.GetTimeRangeFromDashboard(t, dashboard.Data)
-
-		metricReq, err := service.GetMetricRequest(context.Background(), dashboard, publicDashboard, 1, publicDashboardQueryDTO)
-
-		require.NoError(t, err)
-		require.Equal(t, from, metricReq.From)
-		require.Equal(t, to, metricReq.To)
-	})
-}
-
-func TestGetQueryDataResponse(t *testing.T) {
-	sqlStore := sqlstore.InitTestDB(t)
-	dashboardStore := dashboardsDB.ProvideDashboardStore(sqlStore, sqlStore.Cfg, featuremgmt.WithFeatures(), tagimpl.ProvideService(sqlStore, sqlStore.Cfg))
-	publicdashboardStore := database.ProvideStore(sqlStore)
-
-	service := &PublicDashboardServiceImpl{
-		log:                log.New("test.logger"),
-		store:              publicdashboardStore,
-		intervalCalculator: intervalv2.NewCalculator(),
-	}
-
-	publicDashboardQueryDTO := PublicDashboardQueryDTO{
-		IntervalMs:    int64(1),
-		MaxDataPoints: int64(1),
-	}
-
-	t.Run("Returns nil when query is hidden", func(t *testing.T) {
-		hiddenQuery := map[string]interface{}{
-			"datasource": map[string]interface{}{
-				"type": "mysql",
-				"uid":  "ds1",
-			},
-			"hide":  true,
-			"refId": "A",
-		}
-		customPanels := []interface{}{
-			map[string]interface{}{
-				"id": 1,
-				"datasource": map[string]interface{}{
-					"uid": "ds1",
-				},
-				"targets": []interface{}{hiddenQuery},
-			}}
-
-		dashboard := insertTestDashboard(t, dashboardStore, "testDashWithHiddenQuery", 1, 0, true, []map[string]interface{}{}, customPanels)
-		dto := &SavePublicDashboardConfigDTO{
-			DashboardUid: dashboard.Uid,
-			OrgId:        dashboard.OrgId,
-			UserId:       7,
-			PublicDashboard: &PublicDashboard{
-				IsEnabled:    true,
-				DashboardUid: "NOTTHESAME",
-				OrgId:        9999999,
-				TimeSettings: timeSettings,
-			},
-		}
-		pubdashDto, err := service.SavePublicDashboardConfig(context.Background(), SignedInUser, dto)
-		require.NoError(t, err)
-
-		resp, _ := service.GetQueryDataResponse(context.Background(), true, publicDashboardQueryDTO, 1, pubdashDto.AccessToken)
-		require.Nil(t, resp)
-	})
-}
-
-func TestBuildMetricRequest(t *testing.T) {
-	sqlStore := db.InitTestDB(t)
-	dashboardStore := dashboardsDB.ProvideDashboardStore(sqlStore, sqlStore.Cfg, featuremgmt.WithFeatures(), tagimpl.ProvideService(sqlStore, sqlStore.Cfg))
-	publicdashboardStore := database.ProvideStore(sqlStore)
-
-	publicDashboard := insertTestDashboard(t, dashboardStore, "testDashie", 1, 0, true, []map[string]interface{}{}, nil)
-	nonPublicDashboard := insertTestDashboard(t, dashboardStore, "testNonPublicDashie", 1, 0, true, []map[string]interface{}{}, nil)
-	from, to := internal.GetTimeRangeFromDashboard(t, publicDashboard.Data)
-
-	service := &PublicDashboardServiceImpl{
-		log:                log.New("test.logger"),
-		store:              publicdashboardStore,
-		intervalCalculator: intervalv2.NewCalculator(),
-	}
-
-	publicDashboardQueryDTO := PublicDashboardQueryDTO{
-		IntervalMs:    int64(10000000),
-		MaxDataPoints: int64(200),
-	}
-
-	dto := &SavePublicDashboardConfigDTO{
-		DashboardUid: publicDashboard.Uid,
-		OrgId:        publicDashboard.OrgId,
-		PublicDashboard: &PublicDashboard{
-			IsEnabled:    true,
-			DashboardUid: "NOTTHESAME",
-			OrgId:        9999999,
-			TimeSettings: timeSettings,
-		},
-	}
-
-	publicDashboardPD, err := service.SavePublicDashboardConfig(context.Background(), SignedInUser, dto)
-	require.NoError(t, err)
-
-	nonPublicDto := &SavePublicDashboardConfigDTO{
-		DashboardUid: nonPublicDashboard.Uid,
-		OrgId:        nonPublicDashboard.OrgId,
-		PublicDashboard: &PublicDashboard{
-			IsEnabled:    false,
-			DashboardUid: "NOTTHESAME",
-			OrgId:        9999999,
-			TimeSettings: defaultPubdashTimeSettings,
-		},
-	}
-
-	_, err = service.SavePublicDashboardConfig(context.Background(), SignedInUser, nonPublicDto)
-	require.NoError(t, err)
-
-	t.Run("extracts queries from provided dashboard", func(t *testing.T) {
-		reqDTO, err := service.buildMetricRequest(
-			context.Background(),
-			publicDashboard,
-			publicDashboardPD,
-			1,
-			publicDashboardQueryDTO,
-		)
-		require.NoError(t, err)
-
-		require.Equal(t, from, reqDTO.From)
-		require.Equal(t, to, reqDTO.To)
-
-		for i := range reqDTO.Queries {
-			require.Equal(t, publicDashboardQueryDTO.IntervalMs, reqDTO.Queries[i].Get("intervalMs").MustInt64())
-			require.Equal(t, publicDashboardQueryDTO.MaxDataPoints, reqDTO.Queries[i].Get("maxDataPoints").MustInt64())
-		}
-
-		require.Len(t, reqDTO.Queries, 2)
-
-		require.Equal(
-			t,
-			simplejson.NewFromAny(map[string]interface{}{
-				"datasource": map[string]interface{}{
-					"type": "mysql",
-					"uid":  "ds1",
-				},
-				"intervalMs":    int64(10000000),
-				"maxDataPoints": int64(200),
-				"refId":         "A",
-			}),
-			reqDTO.Queries[0],
-		)
-
-		require.Equal(
-			t,
-			simplejson.NewFromAny(map[string]interface{}{
-				"datasource": map[string]interface{}{
-					"type": "prometheus",
-					"uid":  "ds2",
-				},
-				"intervalMs":    int64(10000000),
-				"maxDataPoints": int64(200),
-				"refId":         "B",
-			}),
-			reqDTO.Queries[1],
-		)
-	})
-
-	t.Run("returns an error when panel missing", func(t *testing.T) {
-		_, err := service.buildMetricRequest(
-			context.Background(),
-			publicDashboard,
-			publicDashboardPD,
-			49,
-			publicDashboardQueryDTO,
-		)
-
-		require.ErrorContains(t, err, ErrPublicDashboardPanelNotFound.Reason)
-	})
-
-	t.Run("metric request built without hidden query", func(t *testing.T) {
-		hiddenQuery := map[string]interface{}{
-			"datasource": map[string]interface{}{
-				"type": "mysql",
-				"uid":  "ds1",
-			},
-			"hide":  true,
-			"refId": "A",
-		}
-		nonHiddenQuery := map[string]interface{}{
-			"datasource": map[string]interface{}{
-				"type": "prometheus",
-				"uid":  "ds2",
-			},
-			"refId": "B",
-		}
-
-		customPanels := []interface{}{
-			map[string]interface{}{
-				"id": 1,
-				"datasource": map[string]interface{}{
-					"uid": "ds1",
-				},
-				"targets": []interface{}{hiddenQuery, nonHiddenQuery},
-			}}
-
-		publicDashboard := insertTestDashboard(t, dashboardStore, "testDashWithHiddenQuery", 1, 0, true, []map[string]interface{}{}, customPanels)
-
-		reqDTO, err := service.buildMetricRequest(
-			context.Background(),
-			publicDashboard,
-			publicDashboardPD,
-			1,
-			publicDashboardQueryDTO,
-		)
-		require.NoError(t, err)
-
-		require.Equal(t, from, reqDTO.From)
-		require.Equal(t, to, reqDTO.To)
-
-		for i := range reqDTO.Queries {
-			require.Equal(t, publicDashboardQueryDTO.IntervalMs, reqDTO.Queries[i].Get("intervalMs").MustInt64())
-			require.Equal(t, publicDashboardQueryDTO.MaxDataPoints, reqDTO.Queries[i].Get("maxDataPoints").MustInt64())
-		}
-
-		require.Len(t, reqDTO.Queries, 1)
-
-		require.NotEqual(
-			t,
-			simplejson.NewFromAny(hiddenQuery),
-			reqDTO.Queries[0],
-		)
-
-		require.Equal(
-			t,
-			simplejson.NewFromAny(nonHiddenQuery),
-			reqDTO.Queries[0],
-		)
-	})
-
-	t.Run("metric request built with 0 queries len when all queries are hidden", func(t *testing.T) {
-		customPanels := []interface{}{
-			map[string]interface{}{
-				"id": 1,
-				"datasource": map[string]interface{}{
-					"uid": "ds1",
-				},
-				"targets": []interface{}{map[string]interface{}{
-					"datasource": map[string]interface{}{
-						"type": "mysql",
-						"uid":  "ds1",
-					},
-					"hide":  true,
-					"refId": "A",
-				}, map[string]interface{}{
-					"datasource": map[string]interface{}{
-						"type": "prometheus",
-						"uid":  "ds2",
-					},
-					"hide":  true,
-					"refId": "B",
-				}},
-			}}
-
-		publicDashboard := insertTestDashboard(t, dashboardStore, "testDashWithAllQueriesHidden", 1, 0, true, []map[string]interface{}{}, customPanels)
-
-		reqDTO, err := service.buildMetricRequest(
-			context.Background(),
-			publicDashboard,
-			publicDashboardPD,
-			1,
-			publicDashboardQueryDTO,
-		)
-		require.NoError(t, err)
-
-		require.Equal(t, from, reqDTO.From)
-		require.Equal(t, to, reqDTO.To)
-
-		require.Len(t, reqDTO.Queries, 0)
 	})
 }
